// angular-client/model/fragmentRequest.ts
/**
 * DeepBDE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Entrada para /fragment/. - smiles: SMILES con hidrógenos explícitos - molecule_id: ID de la molécula - bond_idx: Índice del enlace (opcional) - export_smiles: Si True, devuelve lista de SMILES - export_xyz: Si True, devuelve bloque XYZ
 */
export interface FragmentRequest { 
    smiles: string;
    molecule_id: string;
    bond_idx?: number | null;
    export_smiles?: boolean;
    export_xyz?: boolean;
}