// angular-client/model/bDEValues.ts
/**
 * DeepBDE API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Representa los valores de BDE para un enlace. - idx: Índice del enlace - bde: Valor de BDE
 */
export interface BDEValues { 
    idx: number;
    bde: number | null;
}